+++ src/components/TodoList/Item.tsx
@@ -1,21 +1,6 @@
 import {useState} from 'react';
 import * as TodoType from '../../types/TodoList';
-
-interface ButtonProps {
-    name: string;
-    type: 'button' | 'submit';
-    testid: string;
-    handler?: () => void;
-    isDisabled?: boolean;
-}
-
-const Button = ({name, type, testid, handler, isDisabled = false}: ButtonProps) => {
-    return (
-        <button type={type} data-testid={testid} onClick={handler} disabled={isDisabled}>
-            {name}
-        </button>
-    );
-};
+import Button from '../../components/common/Button';
 
 interface ItemProps {
     item: TodoType.Item;
